refactor(Nodes): remove graph from nodes page (and move filter)
Closes DCOS-39457
